Финальный вариант задания

+++ src/Actor.java
@@ -1,6 +1,7 @@
 public class Actor extends Person {
 
-    public Actor(String name, String surname, double height, Gender gender) {
+
+    public Actor(String name, String surname, int height, Gender gender) {
         super(name, surname, height, gender);
     }
