+++ src/Opera.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class Opera extends MusicalShow {
     int choirSize;
 
-    public Opera(ArrayList<Actor> actor, String title, double duration, Director director, String musicAuthor,
+    public Opera(ArrayList<Actor> actor, String title, int duration, Director director, String musicAuthor,
                  String librettoText, int choirSize) {
         super(actor, title, duration, director, musicAuthor, librettoText);
         this.choirSize = choirSize;
